refactor(frontend): extract useCustomMetricDelete hook
Deduplicates the inline delete logic shared by SelectedFieldsSection and
TreeSingleNode, per review.
